backend/server/apps/common/permissions.py
from rest_framework import permissions


class DummyPermission(permissions.BasePermission):
    def has_permission(self, request, view):
        return True

    def has_object_permission(self, request, view, obj):
        return True
