backend/requirements.txt
django==3.0.2
djangorestframework==3.11.0
django-filter==2.2.0
markdown==3.1.1
addict==2.2.1
django-dotenv==1.4.2
django-cors-headers==3.2.1
requests==2.22.0
